pyproject.toml
[project]
name = "nexa-multimodal-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.3",
    "google-genai>=1.70.0",
    "gradio>=6.11.0",
    "jupyterlab>=4.5.6",
    "langchain-text-splitters>=1.1.1",
    "mistralai>=1.0.0",
    "openai>=2.30.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings[dotenv]>=2.13.1",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.9.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "tqdm>=4.67.3",
    "uvicorn>=0.44.0",
]
